+++ server/server/database_manager.h
@@ -9,18 +9,18 @@ public:
 	void CloseODBC_DB();
 
 	bool get_user_info(
-		long long user_id, std::string& out_name,
+		int user_id, std::string& out_name,
 		short& out_x, short& out_y, char& out_dir,
 		short& out_max_hp, short& out_hp, short& out_level, int& out_exp);
 
 	bool insert_user_info(
-		long long user_id,
+		int user_id,
 		const std::string& name,
 		short x, short y, char dir,
 		short max_hp, short hp, short level, int exp);
 
 	bool update_user_info(
-		long long userid, int x, int y,
+		int userid, int x, int y,
 		char dir,
 		int max_hp, int hp,
 		int level, int exp,
